Admin Site Updated

+++ client/src/Dashboard/Admin/Collection/Collection.js
@@ -18,6 +18,10 @@ const Collection = () => {
 
 
     const handleDelete = (leads) => {
+        const confirmDelete = window.confirm('Are you sure you want to delete this item?');
+        if (!confirmDelete) {
+            return;
+        }
 
         fetch(`https://demo-usc-crm-server.vercel.app/delete-collection/${leads._id}`, {
             method: 'DELETE',
