+++ client/src/Dashboard/Admin/Collection/CollectionHead.js
@@ -42,7 +42,10 @@ const CollectionHead = () => {
     }
 
     const handleDelete = (leads) => {
-        console.log(leads);
+        const confirmDelete = window.confirm('Are you sure you want to delete this item?');
+        if (!confirmDelete) {
+            return;
+        }
 
         fetch(`https://demo-usc-crm-server.vercel.app/delete-collection-head/${leads}`, {
             method: 'DELETE',
