+++ client/src/Dashboard/Admin/Settings/CourseSetting.js
@@ -41,7 +41,10 @@ const CourseSetting = () => {
     }
 
     const handleDelete = (leads) => {
-        console.log(leads);
+        const confirmDelete = window.confirm('Are you sure you want to delete this item?');
+        if (!confirmDelete) {
+            return;
+        }
 
         fetch(`https://demo-usc-crm-server.vercel.app/delete-course/${leads}`, {
             method: 'DELETE',
